move type defs
